User mail@example.com as default email
This is just a minor change, but it always triggers me. "example.com" is [reserved by the iana for exactly such usecases](https://www.iana.org/domains/reserved). It will never be taken by someone else. On the other side "mail.com" IS reserved. That'swhy it should be changed.

Note:
There is another place in the code, where this mail is used. I am not sure, if it is also save to change this place as well. I have not tested this change, please double test.
https://github.com/brainstormforce/ultimate-addons-for-gutenberg/blob/b738aa4b75db145e2f601b55c7c28f186fe3fdbe/languages/ultimate-addons-for-gutenberg.pot#L5314